+++ src/wmbus/devices/__init__.py
@@ -1,4 +1,5 @@
 from wmbus.devices.device import Device
 from wmbus.devices.weptech_oms import WeptechOMSv1, WeptechOMSv2
+from wmbus.devices.mock_device import MockDevice
 
-__all__ = ["Device", "WeptechOMSv1", "WeptechOMSv2"]
+__all__ = ["Device", "WeptechOMSv1", "WeptechOMSv2", "MockDevice"]