+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- different tests
+
 ## [0.1.1] - 2019-08-19
 
 ### Added